properly organize classes under SegmentSets module
