+++ lib/emendate/date_segmenter.rb
@@ -10,8 +10,8 @@ module Emendate
 
     def initialize(tokens:, options: {})
       @options = options
-      @working = Emendate::MixedSet.new.copy(tokens)
-      @result = Emendate::MixedSet.new.copy(tokens)
+      @working = Emendate::SegmentSets::MixedSet.new.copy(tokens)
+      @result = Emendate::SegmentSets::MixedSet.new.copy(tokens)
       result.clear
     end
 
@@ -252,7 +252,7 @@ module Emendate
     end
 
     def consume_date_parts
-      pieces = Emendate::MixedSet.new
+      pieces = Emendate::SegmentSets::MixedSet.new
       until working.empty? || current.date_part? == false
         pieces << current
         working.delete(current)
