+++ lib/emendate/date_types/date_type.rb
@@ -15,13 +15,13 @@ module Emendate
       attr_accessor :partial_indicator, :range_switch, :source_tokens
 
       def initialize(**opts)
-        @source_tokens = opts[:children].nil? ? Emendate::MixedSet.new : Emendate::MixedSet.new(opts[:children])
+        @source_tokens = opts[:children].nil? ? Emendate::SegmentSets::MixedSet.new : Emendate::SegmentSets::MixedSet.new(opts[:children])
         @partial_indicator = opts[:partial_indicator]
         @range_switch = opts[:range_switch]
         @certainty = opts[:certainty].nil? ? [] : opts[:certainty]
         @location = location
       end
-
+      
       def date_part?
         true
       end
