+++ lib/emendate/segment/derived_segment.rb
@@ -50,7 +50,7 @@ module Emendate
     end
     
     def set_sources(opts)
-      @sources = Emendate::MixedSet.new
+      @sources = Emendate::SegmentSets::MixedSet.new
       return if opts[:sources].nil?
       return if opts[:sources].empty?
