--- lib/emendate/segment_set/mixed_set.rb
@@ -1,8 +1,0 @@
-# frozen_string_literal: true
-
-require 'emendate/segment_set/segment_set'
-
-module Emendate
-  class MixedSet < Emendate::SegmentSet
-  end
-end
