+++ lib/emendate/segment_sets.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+module Emendate
+  # namespace for SegmentSets
+  module SegmentSets
+  end
+end
