+++ lib/emendate/segment_sets/mixed_set.rb
@@ -1,0 +1,10 @@
+# frozen_string_literal: true
+
+require_relative 'segment_set'
+
+module Emendate
+  module SegmentSets
+    class MixedSet < Emendate::SegmentSets::SegmentSet
+    end
+  end
+end
