+++ lib/emendate/token_collapser.rb
@@ -9,7 +9,7 @@ module Emendate
     attr_reader :result, :options
 
     def initialize(tokens:, options: {})
-      @result = Emendate::TokenSet.new.copy(tokens)
+      @result = Emendate::SegmentSets::TokenSet.new.copy(tokens)
       @options = options
     end
